Added genre `artwork` and Added `sortType` to `queryAudiosFrom` (`IOS`)

+++ on_audio_query/ios/Classes/query/OnGenresQuery.swift
@@ -48,7 +48,7 @@ class OnGenresQuery {
                 var genreData = loadGenreItem(genre: genre)
                 
                 // Count and add the number of songs for every genre.
-                let tmpMediaCount = getMediaCount(type: 0, id: genreData["_id"] as! Int)
+                let tmpMediaCount = getMediaCount(type: 0, id: genreData["_id"] as! UInt64)
                 genreData["num_of_songs"] = tmpMediaCount
                 
                 listOfGenres.append(genreData)
